requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
pytest==7.4.3
pytest-html==4.1.1
pytest-bdd==7.0.1
behave==1.2.6
allure-behave==2.13.2
allure-pytest==2.13.2
python-dotenv==1.0.0
